requirements.txt
# EMERGENCY MINIMAL dependencies for Railway 512MB deployment
torch==2.1.0
# torchvision==0.16.0  # Skip if needed
pinecone==5.0.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pillow==10.0.1
huggingface-hub==0.20.0
numpy==1.24.4
# psutil==5.9.5  # Optional - only if memory allows
# open-clip-torch==2.20.0  # SKIPPED - too heavy for Railway
# timm==0.9.5  # SKIPPED - too heavy for Railway